hadle incorrect path variable

+++ repository/src/main/java/com/epam/esm/mapper/CertificateMapper.java
@@ -18,7 +18,7 @@ public interface CertificateMapper {
     void save(GiftCertificate giftCertificate);
 
     @Delete("DELETE FROM certificates WHERE id = #{id}")
-    int delete(int id);
+    int delete(long id);
 
     @SelectProvider(type = CertificateSqlUtil.class, method = "getByParameter")
     @Results(value = {
